+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/spright-components
 
-This log was last generated on Mon, 22 Jul 2024 15:45:21 GMT and should not be manually modified.
+This log was last generated on Tue, 23 Jul 2024 21:56:35 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.0.10
+
+Tue, 23 Jul 2024 21:56:35 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.1
+
+## 2.0.9
+
+Tue, 23 Jul 2024 20:28:50 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.0
 
 ## 2.0.8